Document: StatusModalNav 삭제
